% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolveSampleSize_trad.R
\name{SolveSampleSize_Single}
\alias{SolveSampleSize_Single}
\title{Calculate the optimal sample sizes for a new two-arm trial}
\usage{
SolveSampleSize_Single(p1, p2, power_level, sig.level = 0.05)
}
\arguments{
\item{p1}{Risk of treatment 1}

\item{p2}{Risk of treatment 2}

\item{power_level}{Power of test we want to obtain}

\item{sig.level}{Significance level, the default value is 0.05}
}
\value{
the optimal sample size for each treatment group
}
\description{
This function calculates the optimal sample size for each treatment group
when 1) the response variable is binary; 2) we analyze it in isolation (w/o the existing network);
}
