% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolveSampleSize_w_nma_equal.R
\name{SolveSampleSize_Withprev_equal}
\alias{SolveSampleSize_Withprev_equal}
\title{Calculate the optimal sample sizes for a new two-arm trial when analyze it with the existing network}
\usage{
SolveSampleSize_Withprev_equal(p1, p2, sigma, power_level, sig.level = 0.05)
}
\arguments{
\item{p1}{Risk of treatment 1}

\item{p2}{Risk of treatment 2}

\item{sigma}{Standard error of the estimated effect size (log odds ratio) between treatment 1 and treatment 2}

\item{power_level}{Power of test we want to obtain}

\item{sig.level}{Significance level, the default value is 0.05}
}
\value{
the total sample size for the new trial
}
\description{
This function calculates the optimal sample size for each treatment group
when planning a new two-arm trial with 1) two arms exist in the existing network;
2) the response variable is binary; 3) we analyze it with the existing network by NMA.
4) the sample size for each treatment group is required to be equal.
}
\examples{
SolveSampleSize_Withprev_equal(p1 = 0.2, p2 = 0.3, sigma = 0.4, power_level = 0.8)
}
