% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolveSampleSize_equal.R
\name{SSS_even}
\alias{SSS_even}
\title{Calculate the optimal sample sizes for a new two-arm trial when analyze it with the existing network}
\usage{
SSS_even(p1, p2, enma_sigma, power.level, sig.level = 0.05, method = "with")
}
\arguments{
\item{p1}{Risk of treatment 1}

\item{p2}{Risk of treatment 2}

\item{power.level}{Power of test we want to obtain}

\item{sig.level}{Significance level, the default value is 0.05}

\item{sigma}{Standard error of the estimated effect size (log odds ratio) between treatment 1 and treatment 2}
}
\value{
the total sample size for the new trial
}
\description{
This function calculates the optimal sample size for each treatment group
when planning a new two-arm trial with 1) two arms exist in the existing network;
2) the response variable is binary; 3) we analyze it with the existing network by NMA.
4) the sample size for each treatment group is required to be equal.
}
